requirements.txt
streamlit==1.32.0
google-generativeai==0.4.0
python-dotenv==1.0.1
Pillow==10.2.0
uuid==1.30
replicate==0.22.0 